Consolidate the API for confirmation dialogs
