+++ wcfsetup/install/files/js/WoltLabSuite/Core/Component/Confirmation.js
@@ -1,26 +1,70 @@
-define(["require", "exports", "./Confirmation/Custom", "./Confirmation/Delete", "./Confirmation/Prefab", "./Confirmation/SoftDelete"], function (require, exports, Custom_1, Delete_1, Prefab_1, SoftDelete_1) {
+define(["require", "exports", "tslib", "./Dialog", "../Language", "../Dom/Util", "./Confirmation/Custom"], function (require, exports, tslib_1, Dialog_1, Language, DomUtil, Custom_1) {
     "use strict";
     Object.defineProperty(exports, "__esModule", { value: true });
     exports.confirmationFactory = void 0;
-    class ConfirmationSetup {
+    Language = tslib_1.__importStar(Language);
+    DomUtil = tslib_1.__importStar(DomUtil);
+    class ConfirmationPrefab {
         custom(question) {
             return new Custom_1.ConfirmationCustom(question);
         }
-        delete(question) {
-            return new Delete_1.ConfirmationDelete(question);
+        async delete(title) {
+            const html = `<p>${Language.get("wcf.dialog.confirmation.cannotBeUndone")}</p>`;
+            const dialog = (0, Dialog_1.dialogFactory)()
+                .fromHtml(html)
+                .asConfirmation({
+                primary: Language.get("wcf.dialog.button.primary.delete"),
+            });
+            const question = Language.get("wcf.dialog.confirmation.delete", { title });
+            dialog.show(question);
+            return new Promise((resolve) => {
+                dialog.addEventListener("primary", () => resolve(true));
+                dialog.addEventListener("cancel", () => resolve(false));
+            });
         }
-        prefab(title) {
-            return new Prefab_1.ConfirmationPrefab(title);
+        async restore(title) {
+            const dialog = (0, Dialog_1.dialogFactory)().withoutContent().asConfirmation();
+            const question = Language.get("wcf.dialog.confirmation.restore", { title });
+            dialog.show(question);
+            return new Promise((resolve) => {
+                dialog.addEventListener("primary", () => resolve(true));
+                dialog.addEventListener("cancel", () => resolve(false));
+            });
         }
-        restore(question) {
-            return this.custom(question);
-        }
-        softDelete(question) {
-            return new SoftDelete_1.ConfirmationSoftDelete(question);
+        async softDelete(title, askForReason) {
+            const dialog = (0, Dialog_1.dialogFactory)().withoutContent().asConfirmation();
+            let reason = undefined;
+            if (askForReason) {
+                const id = DomUtil.getUniqueId();
+                const label = Language.get("wcf.dialog.confirmation.softDelete.reason");
+                const dl = document.createElement("dl");
+                dl.innerHTML = `
+        <dt><label for="${id}">${label}</label></dt>
+        <dd><textarea id="${id}" cols="40" rows="3"></textarea></dd>
+      `;
+                reason = dl.querySelector("textarea");
+                dialog.append(reason);
+            }
+            const question = Language.get("wcf.dialog.confirmation.softDelete", { title });
+            dialog.show(question);
+            return new Promise((resolve) => {
+                dialog.addEventListener("primary", () => {
+                    resolve({
+                        result: true,
+                        reason: reason ? reason.value.trim() : "",
+                    });
+                });
+                dialog.addEventListener("cancel", () => {
+                    resolve({
+                        result: false,
+                        reason: "",
+                    });
+                });
+            });
         }
     }
     function confirmationFactory() {
-        return new ConfirmationSetup();
+        return new ConfirmationPrefab();
     }
     exports.confirmationFactory = confirmationFactory;
 });
